crewai_intelligent_feedback_analysis_action/requirements.txt
# Production requirements with compatible versions for Python 3.10
crewai>=0.86.0
crewai-tools>=0.25.0
langchain>=0.2.0
langchain-openai>=0.1.0
pandas==2.1.4
numpy==1.24.3
streamlit==1.29.0
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
openpyxl==3.1.5
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
openai>=1.40.0